pyproject.toml
[project]
name = "seismic-detection"
version = "1.0.0"
description = "Seismic detection across the Solar System using deep learning - NASA Space Apps 2024"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
    "obspy>=1.4.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "scikit-learn>=1.3.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "flask>=3.0.0",
    "werkzeug>=3.0.0",
    "seaborn>=0.12.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.10.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.25.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
